# apidoc.yaml
openapi: 3.0.0
info:
  title: 'Cafe Manager API'
  description: "API routes for managing cafe's inventory, menu and orders."
  version: 1.0.0
paths:
...
  /api/v1/order:
    post:
      tags:
        - orders
      summary: 'Creates order.'
      operationId: post_app_api_v1_order_saveorder
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OrderRequestDTO'
      responses:
        '200':
          description: 'Order is created successfully.'
          content:
            application/json:
              schema: {  }
              example:
                success: true
    patch:
      tags:
        - orders
      summary: 'Updates order.'
      operationId: patch_app_api_v1_order_updateorder
      parameters:
        -
          name: orderId
          in: query
          description: 'Order ID'
          required: true
          schema:
            type: integer
        -
          name: dish_id
          in: query
          required: false
          schema:
            type: integer
        -
          name: user_id
          in: query
          required: false
          schema:
            type: integer
        -
          name: status
          in: query
          required: false
          schema:
            type: string
        -
          name: is_delivery
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: 'Order is updated successfully.'
          content:
            application/json:
              schema: {  }
              example:
                success: true
  '/api/v1/order/q/{isAsync}':
    post:
      tags:
        - orders
      summary: 'Creates order with queue.'
      operationId: post_app_api_v1_order_saveorderwithqueue
      parameters:
        -
          name: isAsync
          in: path
          required: true
          schema:
            type: string
            pattern: 0|1
            enum:
              - '0'
              - '1'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OrderRequestDTO'
      responses:
        '200':
          description: 'Order is created successfully.'
          content:
            application/json:
              schema: {  }
              example:
                success: true
  '/api/v1/order/by-user/{user_id}':
    get:
      tags:
        - orders
      summary: 'Lists orders by user.'
      operationId: get_app_api_v1_order_getuserorders
      parameters:
        -
          name: user_id
          in: path
          required: true
          schema:
            type: string
            pattern: \d+
      responses:
        '200':
          description: 'Array of user orders is retrieved successfully.'
          content:
            application/json:
              schema:
                properties:
                  orders: { type: array, items: { $ref: '#/components/schemas/OrderResponseDTO' } }
                type: object
  '/api/v1/order/by-dish/{dish_id}':
    get:
      tags:
        - orders
      summary: 'Lists orders by dish.'
      operationId: get_app_api_v1_order_getdishorders
      parameters:
        -
          name: dish_id
          in: path
          required: true
          schema:
            type: string
            pattern: \d+
      responses:
        '200':
          description: 'Array of dish orders is retrieved successfully.'
          content:
            application/json:
              schema:
                properties:
                  orders: { type: array, items: { $ref: '#/components/schemas/OrderResponseDTO' } }
                type: object
  '/api/v1/order/pay/{id}':
    post:
      tags:
        - orders
      summary: 'Performs payment.'
      operationId: post_app_api_v1_order_payorder
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
            pattern: \d+
      responses:
        '200':
          description: 'Order is paid successfully.'
          content:
            application/json:
              schema: {  }
              example:
                success: true
...
  '/api/v1/order/cancel/{id}':
    post:
      tags:
        - orders
      summary: 'Cancels order.'
      operationId: post_app_api_v1_order_cancelorder
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
            pattern: \d+
      responses:
        '200':
          description: 'Order is cancelled successfully.'
          content:
            application/json:
              schema: {  }
              example:
                success: true
  '/api/v1/order/{id}':
    delete:
      tags:
        - orders
      summary: 'Deletes order by ID.'
      operationId: delete_app_api_v1_order_deleteorderbyid
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
            pattern: \d+
      responses:
        '200':
          description: 'Order is deleted successfully.'
          content:
            application/json:
              schema: {  }
              example:
                success: true
  /api/v1/order/stats:
    get:
      tags:
        - orders
      summary: 'Returns total sales per day.'
      operationId: get_app_api_v1_order_getpaidorders
      parameters:
        -
          name: page
          in: query
          description: Page
          schema:
            type: integer
        -
          name: perPage
          in: query
          description: 'Per page'
          schema:
            type: integer
      responses:
        '200':
          description: 'Order stats are retrieved successfully.'
          content:
            application/json:
              schema: {  }
              example:
                orders:
                  - { orderDate: '2024-10-18', total: '330.00' }
  /api/v1/order/get-orders-by-query:
    get:
      tags:
        - orders
      summary: 'Get orders by query.'
      operationId: get_app_api_v1_order_getordersbyquery
      parameters:
        -
          name: query
          in: query
          description: 'Query string'
          schema:
            type: string
        -
          name: page
          in: query
          description: Page
          schema:
            type: integer
        -
          name: perPage
          in: query
          description: 'Per page'
          schema:
            type: integer
      responses:
        '200':
          description: 'Array of orders is retrieved successfully.'
          content:
            application/json:
              schema:
                properties:
                  orders: { type: array, items: { $ref: '#/components/schemas/OrderResponseDTO' } }
                type: object
  /api/v1/order/get-orders-with-aggregation:
    get:
      tags:
        - orders
      summary: 'Get orders with aggregation.'
      operationId: get_app_api_v1_order_getorderswithaggregation
      parameters:
        -
          name: field
          in: query
          description: Field
          schema:
            type: string
      responses:
        '200':
          description: 'Orders grouped by status are retrieved successfully.'
          content:
            application/json:
              schema: {  }
              example:
                orders:
                  doc_count_error_upper_bound: 0
                  sum_other_doc_count: 0
                  buckets: [{ key: created, doc_count: 24 }, { key: paid, doc_count: 9 }, { key: deleted, doc_count: 3 }, { key: cancelled, doc_count: 1 }]
...
components:
  schemas:
...
    OrderRequestDTO:
      properties:
        dish_id:
          type: integer
        user_id:
          type: integer
        status:
          type: string
        is_delivery:
          type: boolean
      type: object
...
    OrderResponseDTO:
      required:
        - id
        - dish
        - user
        - status
        - is_delivery
        - created_at
        - updated_at
      properties:
        id:
          type: integer
        status:
          type: string
        is_delivery:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
      type: object
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
security:
  -
    Bearer: []
...